Move even more, plus cummulative methods
